refactor: 피드백 받기 전 코드 before로 이동
