+++ src/main/java/lottobefore/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lottobefore.domain;
 
 import java.util.List;
