+++ src/main/java/lottobefore/domain/Person.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lottobefore.domain;
 
 import java.util.ArrayList;
 import java.util.HashMap;
