+++ src/test/java/lottobefore/domain/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lottobefore.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
